requirements.txt
# Web Framework and Extensions
Flask
Flask-Cors

# Environment Variables
python-dotenv

# LLM Integration
langchain
langchain-groq

# Utility Packages
requests
python-dateutil
typing-extensions

# Development Tools
black
pylint
mypy
gunicorn

